+++ docs/docusaurus.config.js
@@ -386,8 +386,16 @@ const config = {
             to: "/main/modules/upgrade/",
           },
           {
-            from: ["/main/modules/capability"],
-            to: "/main/ecosystem",
+            from: ["/main/modules/capability", "/main/ecosystem"],
+            to: "/main/modules",
+          },
+          {
+            from: ["/main/spec/circuit-breaker"],
+            to: "/main/modules/circuit",
+          },
+          {
+            from: ["/main/spec/reserve-pool"],
+            to: "/main/modules/distribution",
           },
           {
             from: ["/main/run-node/cosmovisor"],