+++ orm/internal/testutil/testutil.go
@@ -84,8 +84,8 @@ var TestFieldSpecs = []TestFieldSpec{
 			if isNil >= 0.95 { // draw a nil 5% of the time
 				return nil
 			}
-			seconds := rapid.Int64Range(-9999999999, 9999999999).Draw(t, "seconds")
-			nanos := rapid.Int32Range(0, 999999999).Draw(t, "nanos")
+			seconds := rapid.Int64Range(ormfield.TimestampSecondsMin, ormfield.TimestampSecondsMax).Draw(t, "seconds")
+			nanos := rapid.Int32Range(0, ormfield.TimestampNanosMax).Draw(t, "nanos")
 			return (&timestamppb.Timestamp{
 				Seconds: seconds,
 				Nanos:   nanos,
@@ -95,8 +95,15 @@ var TestFieldSpecs = []TestFieldSpec{
 	{
 		"dur",
 		rapid.Custom(func(t *rapid.T) protoreflect.Message {
-			seconds := rapid.Int64Range(0, 315576000000).Draw(t, "seconds")
-			nanos := rapid.Int32Range(0, 999999999).Draw(t, "nanos")
+			isNil := rapid.Float32().Draw(t, "isNil")
+			if isNil >= 0.95 { // draw a nil 5% of the time
+				return nil
+			}
+			seconds := rapid.Int64Range(ormfield.DurationNanosMin, ormfield.DurationNanosMax).Draw(t, "seconds")
+			nanos := rapid.Int32Range(0, ormfield.DurationNanosMax).Draw(t, "nanos")
+			if seconds < 0 {
+				nanos = -nanos
+			}
 			return (&durationpb.Duration{
 				Seconds: seconds,
 				Nanos:   nanos,