+++ x/slashing/keeper/migrations.go
@@ -5,6 +5,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/slashing/exported"
 	v2 "github.com/cosmos/cosmos-sdk/x/slashing/migrations/v2"
 	v3 "github.com/cosmos/cosmos-sdk/x/slashing/migrations/v3"
+	v4 "github.com/cosmos/cosmos-sdk/x/slashing/migrations/v4"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -29,4 +30,11 @@ func (m Migrator) Migrate1to2(ctx sdk.Context) error {
 // module state.
 func (m Migrator) Migrate2to3(ctx sdk.Context) error {
 	return v3.Migrate(ctx, ctx.KVStore(m.keeper.storeKey), m.legacySubspace, m.keeper.cdc)
+}
+
+// Migrate3to4 migrates the x/slashing module state from the consensus
+// version 3 to version 4. Specifically, it migrates the validator missed block
+// bitmap.
+func (m Migrator) Migrate3to4(ctx sdk.Context) error {
+	return v4.Migrate(ctx, m.keeper.cdc, ctx.KVStore(m.keeper.storeKey), m.keeper.GetParams(ctx))
 }